Added Test and refactored gerateChildren
