+++ src/main/scala/model/world/Reproduction.scala
@@ -31,20 +31,18 @@ object Reproduction {
     var childrenGenotypes = List.fill(CHILDREN_FOR_EACH_COUPLE)(PartialGenotype(Map()))
 
     Genes.values.foreach(gk => {
-      val grandmaMomAllele = StandardAllele(mom.genotype(gk).momAllele.kind)
-      val grandpaMomAllele = StandardAllele(mom.genotype(gk).dadAllele.kind)
-      val grandmaDadAllele = StandardAllele(dad.genotype(gk).momAllele.kind)
-      val grandpaDadAllele = StandardAllele(dad.genotype(gk).dadAllele.kind)
-      val genesOfReproduction : List[Gene]= List( Gene(gk, grandmaMomAllele, grandmaDadAllele),
-        Gene(gk, grandpaMomAllele, grandmaDadAllele),
-        Gene(gk, grandmaMomAllele, grandpaDadAllele),
-        Gene(gk, grandpaMomAllele, grandpaDadAllele) )
-      val shuffledGenes = Random.shuffle(genesOfReproduction)
-      childrenGenotypes = (for (i <- 0 until CHILDREN_FOR_EACH_COUPLE) yield childrenGenotypes(i) + shuffledGenes(i)).toList
+      val genesOfReproduction : List[Gene]= Random.shuffle (
+        (for { momAllele <- mom.getStandardAlleles(gk)
+               dadAllele <- dad.getStandardAlleles(gk)
+              } yield Gene(gk, momAllele, dadAllele)).toList)
+
+      childrenGenotypes =
+        (for (i <- 0 until CHILDREN_FOR_EACH_COUPLE)
+          yield childrenGenotypes(i) + genesOfReproduction(i)).toList
 
       mutations filter { _.geneKind == gk } foreach { m =>
           childrenGenotypes = childrenGenotypes(CHILDREN_FOR_EACH_COUPLE - 1) +
-          Gene(m.geneKind, JustMutatedAllele(gk.mutated), JustMutatedAllele(gk.mutated)) :: childrenGenotypes.take(CHILDREN_FOR_EACH_COUPLE - 1)}
+          Gene(gk, JustMutatedAllele(gk.mutated), JustMutatedAllele(gk.mutated)) :: childrenGenotypes.take(CHILDREN_FOR_EACH_COUPLE - 1)}
 
     })
     childrenGenotypes.map(cg => new ChildBunny(CompletedGenotype(cg.genes), Option(mom), Option(dad)))
